+++ app/controllers/application_controller.rb
@@ -1,4 +1,6 @@
 class ApplicationController < ActionController::Base
+  include BaseAuth
+
   before_action :set_current_user
 
   private
