+++ app/controllers/tpi/sectors_controller.rb
@@ -1,5 +1,5 @@
 module TPI
-  class SectorsController < ApplicationController
+  class SectorsController < TPIController
     def index
       @sectors = TPISector.select(:id, :name, :slug).order(:name)
       @companies = Company.select(:id, :name, :slug).limit(50)
