+++ app/controllers/tpi/tpi_controller.rb
@@ -1,0 +1,7 @@
+module TPI
+  class TPIController < ApplicationController
+    include BaseAuth
+
+    layout 'tpi'
+  end
+end
