+++ app/views/layouts/application.html.erb
@@ -9,7 +9,7 @@
     <%= stylesheet_link_tag 'tpi' %>
     <%= stylesheet_link_tag 'application' %>
   </head>
-  
+
   <body>
     <%= yield %>
     <%= javascript_include_tag 'application', 'data-turbolinks-track': 'reload' %>
