+++ app/views/layouts/tpi.html.erb
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>TPI</title>
+    <%= csrf_meta_tags %>
+    <%= csp_meta_tag %>
+
+    <%= javascript_include_tag 'application', 'data-turbolinks-track': 'reload' %>
+    <%= stylesheet_link_tag 'application' %>
+  </head>
+  <body>
+    <main>
+      <div class="container">
+        <%= yield %>
+      </div>
+    </main>
+  </body>
+</html>
